#4 Moved item and recipe databases to database module.
