+++ shopping_list/database/item.py
